Update: Moving files from Internals to Sockets section in Common
