+++ src/libraries/Common/src/System/Net/Sockets/SocketExceptionFactory.Windows.cs
@@ -1,15 +1,14 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Net.Sockets;
-
-namespace System.Net.Internals
+namespace System.Net.Sockets
 {
     internal static partial class SocketExceptionFactory
     {
         public static SocketException CreateSocketException(int socketError, EndPoint? endPoint)
         {
-            return new SocketException(socketError, endPoint?.ToString());
+            // Windows directly maps socketError to native error code.
+            return new SocketException(socketError, CreateMessage(socketError, endPoint));
         }
     }
 }
