+++ src/libraries/Common/src/System/Net/Sockets/SocketExceptionFactory.cs
@@ -1,0 +1,16 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System.Runtime.InteropServices;
+
+namespace System.Net.Sockets
+{
+    internal static partial class SocketExceptionFactory
+    {
+        private static string CreateMessage(int nativeSocketError, EndPoint? endPoint)
+        {
+            string message = Marshal.GetPInvokeErrorMessage(nativeSocketError);
+            return endPoint == null ? message : message + " " + endPoint.ToString();
+        }
+    }
+}
